refactor: move ColumnFamily class testing to test_column_family.py
